Create ATM machine
Create a program that will accept ATM card details such as account number, name, pin and account balance to verify the owner of the atm card. The program should also accept a pin to be entered by the user to verify if it matches the account pin .If it does not match ,the user can repeat  the process again and if it matches, the account balance will be visible on screen and the user can now enter the amount to be transfer or  deposit or withdraw . Afterwards the new account balance should be displayed.